requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Scheduling & Background Tasks
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Development & Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2